Moved wedding info to separate component for rsvp form
